+++ src/main/java/org/thon/transcribestreaming/ByteToAudioEventSubscription.java
@@ -15,7 +15,7 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.amazonaws.transcribestreaming;
+package org.thon.transcribestreaming;
 
 import org.reactivestreams.Subscriber;
 import org.reactivestreams.Subscription;
