+++ src/main/java/org/thon/transcribestreaming/WindowController.java
@@ -15,7 +15,7 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.amazonaws.transcribestreaming;
+package org.thon.transcribestreaming;
 
 import javafx.application.Platform;
 import javafx.geometry.Insets;
